Image File Upload improvements as suggested by Tim: - unique filenames for uploaded image files are now generated using global unique identifiers (simpler than having to normalize the Project.Name) - the file manager now has a constructor which enforces that it get's properly initialized at instantiation. - all state data has been removed from the ImageFileManager class and instead filename, and extension are passed to methods via parameters     (simplifies readability, less chance of adding bugs, and the state doesn't unnecessarily persist after the managers UploadFormFile method has been called) - as necessitated by the above change the WebPath property has been changed to a private helper method that takes in fileName and extensions.

+++ src/CollAction/Helpers/ImageFileManager.cs
@@ -13,39 +13,31 @@ namespace CollAction.Helpers
 {
     public class ImageFileManager
     {
-        private IFormFile _formFile = null;
-        private ImageFile _imageFile = null;
-        private string _fileName = "";
+        private ApplicationDbContext _context;
+        private string _webRoot;
+        private string _webFolder;
 
-        public ApplicationDbContext Context { get; set; }
-
-        public string WebRoot { get; set; }
-
-        public string WebFolder { get; set; }
-        
-        private string FileName { get { return String.Format("{0}{1}", _fileName, Path.GetExtension(_formFile.FileName).ToLower()); } }
-        
-        private string WebPath { get { return Path.Combine(WebFolder, FileName); } }
+        public ImageFileManager(ApplicationDbContext context, string webRoot, string webFolder)
+        {
+            _context = context;
+            _webRoot = webRoot;
+            _webFolder = webFolder;
+        }
 
         public async Task<ImageFile> UploadFormFile(IFormFile formFile, string fileName)
         {
             if (formFile == null) { return null; }
-
-            _formFile = formFile;
-            _fileName = fileName;
-
-            await SaveFileToFileSystem();
-
-            return CreateImageFileModel();
+            string extension = Path.GetExtension(formFile.FileName).ToLower().Substring(1); // Strip the "."
+            await SaveFileToFileSystem(formFile, fileName, extension);
+            return CreateImageFileModel(fileName, extension);
         }
 
-        private async Task SaveFileToFileSystem()
+        private async Task SaveFileToFileSystem(IFormFile formFile, string fileName, string extension)
         {
-            // Save the uploaded image to the file system.
-            var fullPath = Path.Combine(WebRoot, WebPath);
+            var fullPath = Path.Combine(_webRoot, GetWebPath(fileName, extension));
             using (var output = new FileStream(fullPath, FileMode.Create))
             {
-                await _formFile.CopyToAsync(output);
+                await formFile.CopyToAsync(output);
             }
         }
 
@@ -58,41 +50,41 @@ namespace CollAction.Helpers
 
         private void DeleteImageFileFromFileSystem(ImageFile imageFile)
         {
-            var fullPath = Path.Combine(WebRoot, imageFile.Filepath.TrimStart(new char[] { '\\' }));
+            var fullPath = Path.Combine(_webRoot, imageFile.Filepath.TrimStart(new char[] { '\\' }));
             var fileInfo = new FileInfo(fullPath);
-            if (fileInfo != null)
-            {
-                fileInfo.Delete();
-            }
+            if (fileInfo != null) { fileInfo.Delete(); }
         }
 
         private void DeleteImageFileFromModel(ImageFile imageFile)
         {
-            Context.ImageFiles.Remove(imageFile);
+            _context.ImageFiles.Remove(imageFile);
         }
 
-        private ImageFile CreateImageFileModel()
+        private ImageFile CreateImageFileModel(string fileName, string extension)
         {
             ImageFile imageFile = null;
-            var fileName = Path.GetFileName(WebPath);
-            var extension = Path.GetExtension(WebPath).ToLower().Substring(1); // Strip the dot in ".<extension>".
-            var fullPath = Path.Combine(WebRoot, WebPath);
-
+            var webPath = GetWebPath(fileName, extension);
+            var fullPath = Path.Combine(_webRoot, webPath);
             using (var input = File.OpenRead(fullPath))
             {
                 var image = Image.FromStream(input);
                 imageFile = new ImageFile
                 {
                     Name = fileName,
-                    Filepath = String.Format("\\{0}", WebPath),
+                    Filepath = "\\" + webPath,
                     Format = extension,
                     Width = image.Width,
                     Height = image.Height,
                     Date = DateTime.Now
                 };
             }
-
             return imageFile;
         }
+
+        private string GetWebPath(string fileName, string extension)
+        {
+            return Path.Combine(_webFolder, String.Format("{0}.{1}", fileName, extension));
+        }
+
     }
 }
